% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{expand_text}
\alias{expand_text}
\title{Expand Deparsed Code}
\usage{
expand_text(code, data)
}
\arguments{
\item{code}{the deparsed code code matching the watch data}

\item{data}{list watch data simplified with [simplify_data()]}
}
\value{
data frame with the deparsed code repeated for each step of the
  function evaluation, with the line that was just evaluated marked with a
  TRUE in the `highlight` column.
}
\description{
Probably should be made part of [simplify_data()].  Not optimized.
}
\seealso{
[simplify_data()]
}
