% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-sort.R
\name{insert_sort}
\alias{insert_sort}
\title{Implementation of Insertion Sort}
\usage{
insert_sort(x)
}
\arguments{
\item{x}{a vector to sort}
}
\value{
a sorted vector
}
\description{
For use in examples, taken from
\href{#https://en.wikipedia.org/wiki/Insertion_sort}{wikipedia}
}
