% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgwatch-package.R
\docType{package}
\name{pkgwatch-package}
\alias{pkgwatch}
\alias{pkgwatch-package}
\title{pkgwatch: Prevents R packages from being loaded into scope}
\description{
Monitors packages being loaded into the global search scope and
    blocks their additions when appropriate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/pkgwatch}
  \item Report bugs at \url{https://github.com/coatless/pkgwatch/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
