% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watch_data}
\alias{watch_data}
\alias{watch_init}
\alias{capture_data}
\title{Manage Watch Data}
\usage{
watch_data()

watch_init(vars)

capture_data(env, line)
}
\description{
Functions to collect and process watch data.
}
