% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchers.R
\name{watch_pkg}
\alias{watch_pkg}
\alias{unwatch_pkg}
\alias{watchlist}
\title{Add, remove, or see package watches}
\usage{
watch_pkg(pkg, verbose = TRUE)

unwatch_pkg(pkg, verbose = TRUE)

watchlist()
}
\arguments{
\item{pkg}{Name of the Package to actively watch for.}

\item{verbose}{Provide a message indicating addition or removal of watch.}
}
\description{
Setups, destructs, or shows the watch for when a package
is loaded into _R_.
}
\examples{
# Prevent a package from being used.
watch_pkg("pkgname")

# Show active watches
watchlist()

# Allow a package to be used again.
unwatch_pkg("pkgname")
}
