% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watcher-package.R
\docType{package}
\name{watcher-package}
\alias{watcher-package}
\title{watcher: Watch the File System for Changes}
\description{
R binding for 'libfswatch', a filesystem monitoring library. All functions are asynchronous and do not block your session. Set watches on files or directories recursively. Filter by type of event. Log activity, or trigger an R function to run when a specified event occurs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shikokuchuo.net/watcher/}
  \item \url{https://github.com/shikokuchuo/watcher/}
  \item Report bugs at \url{https://github.com/shikokuchuo/watcher/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@shikokuchuo.net} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder]
}

}
\keyword{internal}
