% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchers.R
\name{watch}
\alias{watch}
\alias{unwatch}
\alias{watchlist}
\title{Add, remove, or see package watches}
\usage{
watch(pkg, verbose = TRUE)

unwatch(pkg, verbose = TRUE)

watchlist()
}
\arguments{
\item{pkg}{Name of the Package to actively watch for.}

\item{verbose}{Provide a message indicating addition or removal of watch.}
}
\description{
Setups, destructs, or shows the watch for when a package
is loaded into _R_.
}
\examples{
# Prevent a package from being used.
watch("pkg")

# Show active watches
watchlist()

# Allow a package to be used again.
unwatch("pkg")
}
