% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watchers.R
\name{watch_add}
\alias{watch_add}
\alias{watch_remove}
\alias{watch_active}
\title{Add, remove, or see package watches}
\usage{
watch_add(pkg, verbose = TRUE)

watch_remove(pkg, verbose = TRUE)

watch_active()
}
\arguments{
\item{pkg}{Name of the Package to actively watch for.}

\item{verbose}{Provide a message indicating addition or removal of watch.}
}
\description{
Setups, destructs, or shows the watch for when a package
is loaded into _R_.
}
\examples{
# Prevent a package from being used.
watch_add("pkg")

# Allow a package to be used again.
watch_remove("pkg")
}
