% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watcher.R
\docType{package}
\name{watcher-package}
\alias{watcher-package}
\title{Record Function State During Evaluation}
\description{
A mechanism to record function state during evaluation for later
analysis. Variable values and meta data are recorded for each
‘top-level’ call.
}
\details{
This is package is experimental and rough around the edges, and was
built with expediency rather than robustness as the guiding principle.
The API may change completely in future iterations if there are any.
More likely there will be no future updates to the project, except in
the unlikely case it garners substantial attention.

See \code{vignette('watcher', package='watcher')} for more details.
}
