% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgwatch-package.R
\docType{package}
\name{watcher-package}
\alias{watcher}
\alias{watcher-package}
\title{watcher: Monitors Package Attachment to Session}
\description{
Prevents packages being loaded into the global search scope.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless/watcher}
  \item Report bugs at \url{https://github.com/coatless/watcher/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

}
\keyword{internal}
