% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watcher-package.R
\docType{package}
\name{watcher-package}
\alias{watcher-package}
\title{watcher: Watch the File System for Changes}
\description{
R binding for 'libfswatch', a file system monitoring library. Watch files, or directories recursively, for changes in the background. Log activity, or call an R function, upon every change event.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://watcher.r-lib.org}
  \item \url{https://github.com/r-lib/watcher}
  \item Report bugs at \url{https://github.com/r-lib/watcher/issues}
}

}
\author{
\strong{Maintainer}: Charlie Gao \email{charlie.gao@posit.co} (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
