% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watcher}
\alias{watcher}
\title{Watch a Filesystem Location}
\usage{
watcher(path = getwd(), callback = NULL)
}
\arguments{
\item{path}{Character path to a file, or directory to watch recursively.
Defaults to the current working directory.}

\item{callback}{A function or formula (see \link[rlang:as_function]{rlang::as_function}) - to be
called each time an event is triggered. The default, \code{NULL}, causes event
flag types and paths to be written to \code{stdout} instead.}
}
\value{
A 'Watcher' R6 class object. Start and stop background monitoring
using the \verb{$start()} and \verb{$stop()} methods - these return a logical value
whether or not they have succeeded.
}
\description{
Create a 'Watcher' on a filesystem location to monitor for changes in the
background.
}
\details{
Events are 'bubbled' such that a single change that triggers multiple event
flags will cause the callback to be called only once.

Default latency is 1s.

It is possible to set a watch on a path that does not currently exist, and it
will be monitored once created.
}
\examples{
w <- watcher(tempdir())
isTRUE(w$start())
w
isTRUE(w$stop())
w

}
