% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch.R
\name{watcher}
\alias{watcher}
\title{Watch a Filesystem Location}
\usage{
watcher(path = getwd(), recursive = TRUE, callback = NULL)
}
\arguments{
\item{path}{character path to a file or directory to watch. Defaults to the
current working directory.}

\item{recursive}{logical value, default TRUE, whether to recursively scan
\code{path}, including all subdirectories.}

\item{callback}{(optional) a function (taking no arguments) to be called each
time an event is triggered - requires the \pkg{later} package. The default,
\code{NULL}, causes event paths and types to be written to \code{stdout} instead.}
}
\value{
A 'Watcher' R6 class object. Start and stop background monitoring
using the \verb{$start()} and \verb{$stop()} methods - these return a logical value
whether or not they have succeeded.
}
\description{
Create a 'Watcher' on a filesystem location to monitor for changes in the
background.
}
\details{
A limited subset of filesystem events are watched, namely: Created, Updated,
Removed and Renamed. Default latency is 1s.
}
\examples{
w <- watcher(tempdir())
isTRUE(w$start())
w
isTRUE(w$stop())
w

}
